Fix StoreManager singleton initialization to allow its use when creating the stores.
The singleton instance and variable have to be fully initialized
before we begin to load the stores as they might use methods
referencing the singleton, like [StoreManager globalManager]. Fix this
by moving store creation from [StoreManager -init] to a new method
called when singleton is up and running.

Add an operation queue and an instance method to access to the global
manager. It will be available to stores wanting to move tasks off the
main thread.